Fix hyperlinks that hardcode the web domain
The website is canonically live at https://www.srcf.net, but for
alternate copies of the site (e.g. dev versions), it's a pain to have
URLs here and there hardcoded with the live website's domain that take
the browser to the live site from a dev copy without much warning.  Most
of these cross-site links seem to have emerged from pages that were
originally authored in Markdown elsewhere (e.g. on notes.srcf.net) and
then copied into the site source as HTML without adapting the URLs.

Links are expected to stay within the same copy of the site wherever
possible (or to point at counterparts, such as a dev version of Control
Panel if navigating from a dev version of the main site).  Fix those
links, so that they work as expected.

Note that there are a few special exceptions of services that live under
the main site's domain for historical reasons but are logically separate
from it, such as the web terminal at /terminal/ and
phpMyAdmin/phpPgAdmin at /php{my,pg}admin/.  These are left alone,
partly because they are not really anything to do with the main site
itself and also because dev versions of the site do not offer those
services, given that they're 'mounted' under https://www.srcf.net/ by
special server config.

We often use SSI variables to substitute the actual site base URL in
hyperlinks, which also allows for a copy of the site to be hosted at
not-the-root (e.g. with a base URL of 'https://example.com/srcf-web/')
although I've chosen to use simpler relative URLs where I can, and
absolute URLs where it makes sense -- namely, when accessing the
'/_srcf/' meta-directory that hosts stylesheets and other assets.